b6 +lg2 pre-review

+++ src/main/java/com/tarasiuk/movieland/service/impl/RatingServiceImpl.java
@@ -7,6 +7,8 @@ import com.tarasiuk.movieland.service.RatingService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+
 @Service
 public class RatingServiceImpl implements RatingService {
 
@@ -16,23 +18,31 @@ public class RatingServiceImpl implements RatingService {
 
     @Override
     public void putRating(RatingRequestDTO ratingRequestDTO) {
-        ratingDAO.deleteRating(ratingRequestDTO);
+        if (ratingRequestDTO.getRating() < 1 || ratingRequestDTO.getRating() > 10) {
+            throw new IllegalArgumentException("Rating should be between 1 and 10");
+        }
+
+        for (Rating rating : ratingDAO.getRatingByMovieId(ratingRequestDTO.getMovieId())) {
+            if (rating.getUserId() == ratingRequestDTO.getUserId()) {
+                ratingDAO.deleteRating(ratingRequestDTO);
+            }
+        }
         ratingDAO.addRating(ratingRequestDTO);
     }
 
 
     @Override
     public Rating getRatingById(int ratingId) {
-        return null;
+        return ratingDAO.getRatingById(ratingId);
     }
 
     @Override
-    public Rating getRatingByMovieId(int movieId) {
-        return null;
+    public List<Rating> getRatingByMovieId(int movieId) {
+        return ratingDAO.getRatingByMovieId(movieId);
     }
 
     @Override
-    public Rating getRatingByUserId(int userId) {
-        return null;
+    public List<Rating> getRatingByUserId(int userId) {
+        return ratingDAO.getRatingByUserId(userId);
     }
 }
